Added xml generating scripts for section and courses directories
