+++ src/components/generateCompletionXml.js
@@ -1,7 +1,9 @@
 const fs = require("fs");
 const path = require("path");
 const xmlbuilder = require("xmlbuilder");
-
+ 
+// Generates completion.xml file inside 'output' directory
+// output\completion.xml
 function generateCompletionXml(outputDir) {
   const completionXml = xmlbuilder
     .create("course_completion", { encoding: "UTF-8" })
