+++ src/components/generateFilesXml.js
@@ -2,6 +2,8 @@ const fs = require('fs');
 const path = require('path');
 const xmlbuilder = require('xmlbuilder');
 
+// Generates files.xml file inside 'output' directory
+// output\files.xml
 function generateFilesXml(outputDir) {
     const fileXml = xmlbuilder.create('files', { encoding: 'UTF-8' })
          .ele('file',  { id: '' })
