+++ src/components/generateGradehistoryXml.js
@@ -2,6 +2,8 @@ const fs = require('fs');
 const path = require('path');
 const xmlbuilder = require('xmlbuilder');
 
+// Generates grade_history.xml file inside 'output' directory
+// output\grade_history.xml
 function generateGradehistoryXml(outputDir) {
     const gradehistoryXml = xmlbuilder.create('grade_history', { encoding: 'UTF-8' })
     .ele('grade_grades', '').up()
