+++ src/components/generateGroupsXml.js
@@ -2,6 +2,8 @@ const fs = require('fs');
 const path = require('path');
 const xmlbuilder = require('xmlbuilder');
 
+// Generates groups.xml file inside 'output' directory
+// output\groups.xml
 function generateGroupsXml(outputDir) {
     const groupsXml = xmlbuilder.create('groups', { encoding: 'UTF-8' })
         .ele('group', { id: '' })
