+++ src/components/generateOutcomesXml.js
@@ -2,6 +2,8 @@ const fs = require('fs');
 const path = require('path');
 const xmlbuilder = require('xmlbuilder');
 
+// Generates outcomes.xml file inside 'output' directory
+// output\outcome.xml
 function generateOutcomesXml(outputDir) {
     const outcomesXml = xmlbuilder.create('outcomes_definition', { encoding: 'UTF-8' })
     .end({ pretty: true });
