+++ src/components/generateQuestionsXml.js
@@ -2,6 +2,8 @@ const fs = require('fs');
 const path = require('path');
 const xmlbuilder = require('xmlbuilder');
 
+// Generates questions.xml file inside 'output' directory
+// output\questions.xml
 function generateQuestionsXml(outputDir) {
     const questionsXml = xmlbuilder.create('question_categories', { encoding: 'UTF-8' })
         .ele('question_category', { id: '' })
